Major Code Clean Up  ( Beta !)

+++ FastLog.Net/Helpers/FileHelper.cs
@@ -30,7 +30,7 @@ namespace FastLog.Helpers
 
                 if (string.IsNullOrWhiteSpace(fileName))
                 {
-                    return (short)0;
+                    return 0;
                 }
                 else
                 {
